+++ evaluators/contrib/defenseclaw/src/agent_control_evaluator_defenseclaw/opa_policy/__init__.py
@@ -1,0 +1,6 @@
+"""DefenseClaw OPA-policy evaluator exports."""
+
+from .config import DefenseClawOpaPolicyConfig, OpaPolicy
+from .evaluator import DefenseClawOpaPolicyEvaluator
+
+__all__ = ["DefenseClawOpaPolicyConfig", "DefenseClawOpaPolicyEvaluator", "OpaPolicy"]