+++ evaluators/contrib/defenseclaw/src/agent_control_evaluator_defenseclaw/rule_pack/__init__.py
@@ -1,0 +1,11 @@
+"""DefenseClaw rule-pack evaluator exports."""
+
+from .config import DefenseClawRulePackConfig, RuleConfig, RulePack
+from .evaluator import DefenseClawRulePackEvaluator
+
+__all__ = [
+    "DefenseClawRulePackConfig",
+    "DefenseClawRulePackEvaluator",
+    "RuleConfig",
+    "RulePack",
+]